Make updates to check_coverage, helpers, and sum_stats_transform to ensure check_coverage runs as expected. Update bash scripts accordingly. Fix main function logic for analyze_colocalization and start bash script to run.

+++ src/helpers.py
@@ -98,7 +98,7 @@ def add_variant_id(
     all_present = pos.notna() & ref.notna() & alt.notna()
     df[variant_id_key] = np.where(
         all_present,
-        chr + ":" + pos.astype(str) + ":" + ref.astype(str) + ":" + alt.astype(str),
+        chr + ":" + pos.astype(int).astype(str) + ":" + ref.astype(str) + ":" + alt.astype(str),
         "NA"
     )
